Move FeesDiscount component to separated file
